+++ packages/backend/src/queryBuilder.ts
@@ -1,90 +1,19 @@
 import {
-    Dimension,
-    FilterGroup,
     Explore,
-    ExploreJoin,
     fieldId,
-    FilterGroupOperator,
-    Metric,
-    MetricQuery,
-    StringFilter,
-    StringDimension,
-    NumberDimension,
-    NumberFilter,
-    getDimensions,
+    FieldId,
     fieldIdFromFilterGroup,
-    FieldId, getMetrics, Table,
+    FilterGroup,
+    FilterGroupOperator,
+    getDimensions,
+    getMetrics,
+    MetricQuery,
+    NumberFilter,
+    StringFilter,
 } from "common";
 
 
-const lightdashVariablePattern = /\$\{([a-zA-Z0-9_.]+)\}/g
-
-const renderDimensionReference = (ref: string, tables: Record<string, Table>, currentTable: string): string => {
-    // Reference to current table
-    if (ref === 'TABLE') {
-        return currentTable
-    }
-    // Reference to another dimension
-    const split = ref.split('.')
-    if (split.length > 2) {
-        throw new Error(`Model ${currentTable} has an illegal dimension reference: \${${ref}}`)
-    }
-    const refTable = split.length === 1 ? currentTable : split[0]
-    const refName = split.length === 1 ? split[0] : split[1]
-    const dimension = tables[refTable]?.dimensions[refName]
-    if (dimension === undefined)
-        throw Error(`Model ${currentTable} has a dimension reference: \${${ref}} which matches no dimension`)
-    return `(${renderDimensionSql(dimension, tables)})`
-}
-
-const renderMetricReference = (ref: string, explore: Explore, currentTable: string): string => {
-    // Reference to current table
-    if (ref === 'TABLE') {
-        return currentTable
-    }
-    // Reference to another dimension
-    const split = ref.split('.')
-    if (split.length > 2) {
-        throw new Error(`Model ${currentTable} has an illegal metric reference: \${${ref}}`)
-    }
-    const refTable = split.length === 1 ? currentTable : split[0]
-    const refName = split.length === 1 ? split[0] : split[1]
-    const metric = explore.tables[refTable]?.metrics[refName]
-    if (metric === undefined)
-        throw Error(`Model ${currentTable} has a metric reference: \${${ref}} which matches no metric`)
-    return `(${renderMetricSql(metric, explore)})`
-}
-
-const renderMetricSql = (metric: Metric, explore: Explore): string => {
-    // Metric might have references to other dimensions
-    const renderedSql = metric.sql.replace(lightdashVariablePattern, (_, p1) => renderDimensionReference(p1, explore.tables, metric.table))
-    const metricType = metric.type
-    switch(metricType) {
-        case "average": return `AVG(${renderedSql})`
-        case "count":   return `COUNT(${renderedSql})`
-        case "count_distinct": return `COUNT(DISTINCT ${renderedSql})`
-        case "max": return `MAX(${renderedSql})`
-        case "min": return `MIN(${renderedSql})`
-        case "sum": return `SUM(${renderedSql})`
-        default:
-            const nope: never = metricType
-            throw Error(`No SQL render function implemented for metric with type ${metric.type}`)
-    }
-}
-
-
-const renderDimensionSql = (dimension: Dimension, tables: Record<string, Table>): string => {
-    // Dimension might have references to other dimensions
-    return dimension.sql.replace(lightdashVariablePattern, (_, p1) => renderDimensionReference(p1, tables, dimension.table))
-}
-
-export const renderExploreJoinSql = (table: string, sqlOn: string, tables: Record<string, Table>): string => {
-    // Sql join contains references to dimensions
-    return sqlOn.replace(lightdashVariablePattern, (_, p1) => renderDimensionReference(p1, tables, table))
-}
-
-const renderStringFilterSql = (dimension: StringDimension, filter: StringFilter, explore: Explore): string => {
-    const dimensionSql = renderDimensionSql(dimension, explore.tables)
+const renderStringFilterSql = (dimensionSql: string, filter: StringFilter): string => {
     const filterType = filter.operator
     switch (filter.operator) {
         case "equals":
@@ -99,12 +28,11 @@ const renderStringFilterSql = (dimension: StringDimension, filter: StringFilter,
             return `(${dimensionSql}) LIKE '${filter.value}%'`
         default:
             const nope: never = filter
-            throw Error(`No function implemented to render sql for filter type ${filterType} on dimension type ${dimension.type}`)
+            throw Error(`No function implemented to render sql for filter type ${filterType} on dimension of number type`)
     }
 }
 
-const renderNumberFilterSql = (dimension: NumberDimension, filter: NumberFilter, explore: Explore): string => {
-    const dimensionSql = renderDimensionSql(dimension, explore.tables)
+const renderNumberFilterSql = (dimensionSql: string, filter: NumberFilter): string => {
     const filterType = filter.operator
     switch (filter.operator) {
         case "equals":
@@ -121,7 +49,7 @@ const renderNumberFilterSql = (dimension: NumberDimension, filter: NumberFilter,
             return `(${dimensionSql}) < ${filter.value}`
         default:
             const nope: never = filter
-            throw Error(`No function implemented to render sql for filter type ${filterType} on dimension type ${dimension.type}`)
+            throw Error(`No function implemented to render sql for filter type ${filterType} on dimension of string type`)
     }
 }
 
@@ -133,12 +61,12 @@ const renderFilterGroupSql = (filterGroup: FilterGroup, explore: Explore): strin
     switch (filterGroup.type) {
         case "string": {
             if (dimension?.type === 'string')
-                return filterGroup.filters.map(filter => renderStringFilterSql(dimension, filter, explore)).join(`\n   ${operator} `)
+                return filterGroup.filters.map(filter => renderStringFilterSql(dimension.compiledSql, filter)).join(`\n   ${operator} `)
             throw new Error(`StringFilterGroup has a reference to an unknown string field ${fieldIdFromFilterGroup(filterGroup)}`)
         }
         case "number":
             if (dimension?.type === 'number')
-                return filterGroup.filters.map(filter => renderNumberFilterSql(dimension, filter, explore)).join(`\n   ${operator} `)
+                return filterGroup.filters.map(filter => renderNumberFilterSql(dimension.compiledSql, filter)).join(`\n   ${operator} `)
             throw new Error(`NumberFilterGroup has a reference to an unknown number field ${fieldIdFromFilterGroup(filterGroup)}`)
         default:
             const nope: never = filterGroup
@@ -183,19 +111,19 @@ export const buildQuery = ({ explore, metricQuery }: BuildQueryProps) => {
     const sqlJoins = explore.joinedTables.map(join => {
         const joinTable = explore.tables[join.table].sqlTable
         const alias = join.table
-        return `LEFT JOIN ${joinTable} AS ${alias}\n  ON ${join.sqlOn}`
+        return `LEFT JOIN ${joinTable} AS ${alias}\n  ON ${join.compiledSqlOn}`
     }).join('\n')
 
     const dimensionSelects = dimensions.map(field => {
         const alias = field
         const dimension = getDimensionFromId(field, explore)
-        return `  ${renderDimensionSql(dimension, explore.tables)} AS ${q}${alias}${q}`
+        return `  ${dimension.compiledSql} AS ${q}${alias}${q}`
     })
 
     const metricSelects = metrics.map(field => {
         const alias = field
         const metric = getMetricFromId(field, explore)
-        return `  ${renderMetricSql(metric, explore)} AS ${q}${alias}${q}`
+        return `  ${metric.compiledSql} AS ${q}${alias}${q}`
     })
 
     const sqlSelect = `SELECT\n${[...dimensionSelects, ...metricSelects].join(',\n')}`